requirements.txt
# Core ML/DL dependencies
torch==2.8.0+cu128
torchvision>=0.19.0
torchaudio>=2.4.0
pytorch-lightning==2.5.0.post0

# Configuration and data handling
omegaconf==2.3.0
datasets==3.3.2
transformers>=4.21.0
accelerate>=0.20.0

# Image processing and utilities  
Pillow>=8.0.0
numpy>=1.21.0
scipy>=1.7.0

# Visualization and logging
matplotlib>=3.5.0
seaborn>=0.11.0
rich>=13.0.0
tensorboard>=2.8.0
wandb>=0.12.0

# Additional ML utilities
timm>=0.6.0
einops>=0.6.0
hydra-core>=1.2.0

# Development and utilities
tqdm>=4.64.0
psutil>=5.8.0